requirements.txt
bcrypt==4.1.2
black==22.12.0
click==8.1.3
colorama==0.4.6
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pathspec==0.10.3
pillow==10.2.0
platformdirs==2.6.2
tk==0.1.0
tomli==2.0.1
typing_extensions==4.4.0
